efficientdet/requirements.txt
numpy==1.26.4
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0+cu118
timm==1.0.22
effdet==0.4.1
pillow==11.3.0
matplotlib==3.10.7
onnx==1.16.1
onnxruntime==1.20.1
protobuf==4.25.3
pycocotools==2.0.8
PyYAML==6.0.1
typing_extensions==4.10.0
tqdm==4.66.1
